requirements.txt
numpy==1.17.2
opencv-contrib-python==3.4.2.17	
opencv-python==3.4.2.17	
Pillow==6.1.0	
tensorflow==1.13.1		
